- Rename story_draft to story_drafts folder
